html/editing/dnd/drop/events-textarea.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Selection drag and drop: events for &lt;input></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div><input id=a value="Drag me"></div>
<div><textarea id=b placeholder="...to here"></textarea></div>
<script>
setup({ explicit_timeout: true, single_test: true });
const a = document.getElementById('a');
const b = document.getElementById('b');
const actualEvents = [];
const expectedEvents = ['drop', 'beforeinput', 'textInput', 'input'];
for (const eventType of expectedEvents) {
  b.addEventListener(eventType, e => {
    actualEvents.push(e.type);
    if (actualEvents.length === expectedEvents.length) {
      assert_array_equals(actualEvents, expectedEvents);
      done();
    }
  });
}
a.select();
</script>
